Convert ParticipantIdentifier.Value to Guid. Add index.

+++ Dynamo.Stream.Handler/src/Dynamo.Stream.Handler/Handlers/StreamHandler.cs
@@ -2,8 +2,8 @@ using Amazon.DynamoDBv2;
 using Amazon.DynamoDBv2.Model;
 using Amazon.Lambda.DynamoDBEvents;
 using Dynamo.Stream.Handler.Entities;
-using Dynamo.Stream.Handler.Extensions;
 using Dynamo.Stream.Handler.Mappers;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using static Amazon.Lambda.DynamoDBEvents.StreamsEventResponse;
 
@@ -98,10 +98,29 @@ public class StreamHandler : IStreamHandler
         await InsertAsync(record.Dynamodb.NewImage, cancellationToken);
     }
 
+    private async Task<Participant> InsertAsync(Dictionary<string, AttributeValue> image, CancellationToken cancellationToken)
+    {
+        var identifiers = _participantMapper.ExtractIdentifiers(image);
+
+        var targetParticipant = await _dbContext.GetParticipantByLinkedIdentifiers(identifiers)
+                                                .ForUpdate()
+                                                .SingleOrDefaultAsync(cancellationToken);
+
+        if (targetParticipant == null)
+        {
+            // No linked participant exists, create a new one.
+            targetParticipant = _dbContext.Participants.Add(new Participant()).Entity;
+        }
+
+        return _participantMapper.Map(image, targetParticipant);
+    }
+
     private async Task ProcessModifyAsync(DynamoDBEvent.DynamodbStreamRecord record, CancellationToken cancellationToken)
     {
-        var pk = record.Dynamodb.NewImage.PK();
-        var participant = await _dbContext.GetParticipantByPkAsync(pk, cancellationToken);
+        var identifiers = _participantMapper.ExtractIdentifiers(record.Dynamodb.NewImage);
+        var participant = await _dbContext.GetParticipantByLinkedIdentifiers(identifiers)
+                                                .ForUpdate()
+                                                .SingleOrDefaultAsync(cancellationToken);
 
         if (participant == null)
         {
@@ -113,16 +132,26 @@ public class StreamHandler : IStreamHandler
 
     private async Task ProcessRemoveAsync(DynamoDBEvent.DynamodbStreamRecord record, CancellationToken cancellationToken)
     {
-        var pk = record.Dynamodb.OldImage.PK();
-        var participant = await _dbContext.GetParticipantByPkAsync(pk, cancellationToken);
+        var identifiers = _participantMapper.ExtractIdentifiers(record.Dynamodb.OldImage);
+        var participant = await _dbContext.GetParticipantByLinkedIdentifiers(identifiers)
+                                          .Include(x => x.ParticipantIdentifiers)
+                                          .SingleOrDefaultAsync(cancellationToken);
 
         if (participant == null)
         {
             participant = await InsertAsync(record.Dynamodb.OldImage, cancellationToken);
         }
 
-        _dbContext.Participants.Remove(participant);
-    }
+        // TODO: are we removing the Participant here, or just the ParticipantIdentifer?
+        // Only delete the Participant if all participant identifiers have also been deleted.
+        var idsToRemove = participant.ParticipantIdentifiers.Where(x => identifiers.Select(y => y.Value).Contains(x.Value));
+
+        _dbContext.ParticipantIdentifiers.RemoveRange(idsToRemove);
+
+        if (participant.ParticipantIdentifiers.Except(idsToRemove).All(x => x.IsDeleted))
+        {
+            _dbContext.Participants.Remove(participant);
+        }
 
     private async Task<Participant> InsertAsync(Dictionary<string, AttributeValue> image, CancellationToken cancellationToken)
     {
@@ -131,7 +160,5 @@ public class StreamHandler : IStreamHandler
         if (targetParticipant == null)
         {
             targetParticipant = _dbContext.Participants.Add(new Participant()).Entity;
-        }
-        return _participantMapper.Map(image, targetParticipant);
     }
 }
